requirements.txt
aws-cdk-lib==2.233.0
constructs==10.4.4
cdk-nag==2.37.55
requests==2.32.5
boto3>=1.42.18